Simplify structure and use new App component for SSoT
To avoid having to load in Redux for an API-powered app, this will be using the App component to handle all logic and Ajax queries. Code splitting has also temporarily been removed to hammer out the routing bugs, and make sure it works as a single bundle. Code will be re-split at the end where it’s beneficial.